% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmn.R
\docType{package}
\name{bpmn-package}
\alias{bpmn-package}
\title{bpmn: display BPMN diagrams}
\description{
An R interface to bpmn-js library \url{https://bpmn.io/toolkit/bpmn-js/}
}
\section{More examples}{


  See \url{https://bergant.github.io/bpmn} for more usage examples.
}

\section{Support}{


  Use \url{https://github.com/bergant/bpmn/issues} for bug reports
}

\examples{
\dontrun{

bpmn(bpmn_xml)

}

}
