% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmn.R
\name{bpmn}
\alias{bpmn}
\title{BPMN diagram}
\usage{
bpmn(process, overlays = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{process}{A file name or xml document or string in BPMN XML format}

\item{overlays}{A list of elements to be added to the diagram's existing
elements. Use overlay function to create a proper structure with content
and relative position.}

\item{width}{Fixed width for widget (in css units). The default is NULL,
which results in intelligent automatic sizing based on the widget's
container.}

\item{height}{Fixed height for widget (in css units). The default is NULL,
which results in intelligent automatic sizing based on the widget's
container.}

\item{elementId}{element id}
}
\description{
Display BPMN diagram based on BPMN definition in XML format
}
