% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmn.R
\name{marker}
\alias{marker}
\title{Marker}
\usage{
marker(element, class_name = "highlight")
}
\arguments{
\item{element}{The bpmn element to which the overlay will be attached}

\item{class_name}{HTML class name to use for highlighting element}
}
\value{
A marker object
}
\description{
Create a marker object
}
