% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmn.R
\name{overlay}
\alias{overlay}
\title{Overlay}
\usage{
overlay(element, html, type = "note", left = NULL, right = NULL,
  top = NULL, bottom = NULL, min_zoom = NULL, max_zoom = NULL)
}
\arguments{
\item{element}{The bpmn element to which the overlay will be attached}

\item{html}{HTML element to use as an overlay to use as an overlay}

\item{type}{Optional type to assign to the overlay}

\item{left}{Where to attach the overlay, relative to element bbox left
attachment}

\item{right}{Where to attach the overlay, relative to element bbox right
attachment}

\item{top}{Where to attach the overlay, relative to element bbox top
attachment}

\item{bottom}{where to attach the overlay, relative to element bbox bottom
attachment}

\item{min_zoom}{minimal zoom level to show the overlay}

\item{max_zoom}{maximum zoom level to show the overlay}
}
\value{
An overlay object
}
\description{
A list of overlays can be added to existing elements in the diagram. See
\code{overlays} argument in \code{\link{bpmn}} function. Use this structure
to create correct overlay structure.
}
