% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmn.R
\name{bpmn}
\alias{bpmn}
\title{BPMN diagram}
\usage{
bpmn(bpmn_xml, overlays = NULL, markers = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{bpmn_xml}{A file name or xml document or string in BPMN XML format}

\item{overlays}{A list of elements to be added to the diagram's existing
elements. Use overlay function to create an overlay object with content
and relative position.}

\item{markers}{A list of markersto highlight existing elements. Use marker
function to create a marker object with target elements.}

\item{width}{Fixed width for widget (in css units). The default is NULL,
which results in intelligent automatic sizing based on the widget's
container.}

\item{height}{Fixed height for widget (in css units). The default is NULL,
which results in intelligent automatic sizing based on the widget's
container.}

\item{elementId}{element id}
}
\description{
Display BPMN diagram based on BPMN definition in XML format
}
